Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> What's the motivation for pulling this out of resource provider?

There is no burning reason (this is more of "clean out the deferred Image_Gpu CL so it is rational" CL than anything else. Since the GrDrawOpAtlas (at least for the time being) takes a GrContext, it seemed odd for this to be on the GrResourceProvider rather than a utility.

Patch-set: 3
Reviewer: Robert Phillips <5230@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0