Update patch set 3
Patch Set 3:

(1 comment)

I'm not totally opposed to it but I don't love the fGpu switch thing. Are there other ownership models that could work? Do you think we'll reuse these? Could they be uniquely owned and waiting destroys it? Or could we put it in some freeze dried state like we're doing with the texture innards and rehydrate it on the receiving side?

Patch-set: 3