Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> (1 comment)
> 
> I'm not totally opposed to it but I don't love the fGpu switch thing. Are there other ownership models that could work? Do you think we'll reuse these? Could they be uniquely owned and waiting destroys it? Or could we put it in some freeze dried state like we're doing with the texture innards and rehydrate it on the receiving side?

Just to add tracking of our offline conversation, in general the semaphores cannot be uniquely owned (in future can exists in multiple ops at once), and can't just be destroyed after wait (in the Vulkan use case of exporting the VkSemaphore to client to wait, we want to destroy after we've signaled).

For now, we will keep this model of passing the same object between two GrContexts since adding code to rip out the semaphore and rebuild on the other context doesn't seem to buy us much. If we ever find needs where we are cross process or other situations where we can't share the object we will revisit this.

Patch-set: 4
Reviewer: Greg Daniel <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>