Update patch set 4
Patch Set 4:

How about we just start by generating new images in Gold and see if we can handle it?  Presumably whatever configs we set up (just srgbnl-gpu? or also srgbnl-8888?) are the most important things to triage after 8888 and gpu, and will soon pass them in importance.  We're already triaging things like 565 that look a bit different from 8888.  If we find it's too much burden to triage, we can set an ignore on them, then figure out a better plan.

Patch-set: 4