Update patch set 1
Patch Set 1:

I think because we're using a lot of the same tools this can be confusing.  Please let me know if this makes sense.

For our local testing, we use GN to describe how to build libskia and all our test apps.  Chrome, Fuchsia, Android, and other clients derive their ways of building libskia from that same GN source of truth.  (Google3 is a little weird.)  Some of those clients use GN to do their builds.  We don't expect anyone but Skia devs builds any of our test apps like DM.

If we were trying to port Skia to, say, Windows on ARMv8, we'd want to get access to a machine running Windows on ARMv8, and a toolchain that can compile for Windows on ARMv8 (not necessarily self hosted, but that's a fine option).  We would then point our GN environment at that toolchain, build libskia and DM, then ship all that over to the machine that can run it.

If we want to run DM on Fuchsia, we'd want access to a machine running Fuchsia on whatever architecture, and a toolchain that can compile for Fuchsia on that machine's architecture.  We would then point our GN environment at that toolchain, build libskia and DM, then ship all that over to the machine that can run it.

I wouldn't think we'd ever want to be aware that Fuchsia is using GN or Skia internally in order to build DM for Fuchsia.

Patch-set: 1
CC: Mike Klein <5010@5086850b-e599-37d2-8ec8-ff16aeccbf02>