Update patch set 3
Patch Set 3:

Pretty much the same speed.  I guess it's limited by the load, not creating the mask?  Either version of this CL really slims the loads and stores on the profile... might be not much left to cut?

Still, readability and code size say this is the way to go.  :D

Patch-set: 3
Reviewer: Mike Klein <5010@5086850b-e599-37d2-8ec8-ff16aeccbf02>