Strengthen is_official_build, update docs.
This makes is_official_build turn off all development targets and
features in Skia, including building third-party dependencies from
source.

This makes gn_to_bp.py a little simpler to write.  In spirit, Android
builds are official Skia builds that also build DM and nanobench.

It seems that SkJumper is (unintentionally) enabled on Android.
Switching to an is_official_build would have disabled that.  But
seeing as it seems to be doing fine, I've now explicitly enabled it.

So, in the end, no changes to Android.bp or its SkUserConfig.h.

Change-Id: Id06e53268a5caf55c6046ada354a0863c3031c73