Update patch set 2
Patch Set 2:

> Patch Set 2:
> But, that uses the heap, not the stack.

The various AdditiveBlitters are very small -- it's aaa_fill_path with SK_ALWAYS_INLINE that causes the large stack frame.

My C++ skills aren't good enough to write this so that the AdditiveBlitters are stack-allocated and there is only one call to aaa_fill_path. It probably involves a goto.

Patch-set: 2