Update patch set 5
Patch Set 5: Code-Review+1

I'm sure this is fine. My hesitation is two-fold
Earlier this week I tried to run extensive pathops tests on Ubuntu (i.e., add -V -x to the dm command line) and was unable to do so because it appeared to hang all processors in string building (outside of the tests themselves). This gives me less confidence that I know how to test this change.
My other concern is that while I'd like to remove the multitude of allocators and have a single allocator shared through a thread local storage pointer, the last time I tried to profile things I didn't see the allocator show up, making me question spending time on it.
So, I assume you're doing this so we can get rid of SkChunkAlloc and have one less thing to maintain. If that be the case, go for it.

Patch-set: 5
Reviewer: Cary Clark <5632@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1