Support Android NDK r14.
simpleperf has moved a little.  To keep both r13 and r14 working, remove
the copy step... it seems more nice-to-have than essential.

Explicitly disable the Clang integrated assembler on MIPS.  It has only
understood the 'usw' (unaligned store) pseudoinstruction for about 4
months, and that won't be part of a Clang release until the upcoming
Clang 4.0.  Must be -no-integrated-as was the default in r13.

All six platforms build on my laptop with r14 now.

Will follow up with bot updates... want to land this first to double
check the existing r13 the bots are using is okay with these changes.

Change-Id: Ic00981d534613ebab8f6e7646b967493728b91d9