Update patch set 4
Patch Set 4:

sorry for my indecision on this, but many people use the SDK manager to download the NDK and it has not yet been updated to r14.  Since I'm one of the few people that uses simpleperf I'll do the copy for now so that we don't inconvenience everyone.

Patch-set: 4