Update patch set 27
Patch Set 27:

> Patch Set 27: Code-Review+1
> 
> lgtm if it lgtc

Just wanted to resolve that last issue of when to allocate when there are lots of atlases.

If we decide we don't care about memory yet and we want to draw them all upfront, that's fine too. I just want to decide on a resolution.

Patch-set: 27
Reviewer: Chris Dalton <5633@5086850b-e599-37d2-8ec8-ff16aeccbf02>