Update patch set 1
Patch Set 1: Code-Review+1

I'm fine with this, but it often seems questionable to use detachAsData. In particular, Chromium doesn't need to linearize the data to write to a file, and when it does want to liniearize into shared memory it already has its own memory. It seems like it would be nicer if there were a detachToBuffer which just liniearized into a buffer and then users could create their own SkData and get the writable buffer out of it and pass that to detachToBuffer.

Patch-set: 1
Reviewer: Ben Wagner <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1