Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Nice change, great to see more DWrite improvements. Thank you! This seems to be controlled by gasp as well, would you recommend waiting for this to roll before I do the Windows rebaselines?
> 
> In earlier patch sets the changes here were dependent a bit on gasp, but they aren't now (because we're not using it as often). So I don't think this affects the gasp work.

Could you explain a little more what individual changes this code does. I don't yet understand it completely. One change seems to be that "native" grayscale antialiasing is used when available on the newer DWrite interface, instead of grayscaling the RGB antialiasing. 

What do you mean by "because we're not using it as often"? 

Is it correct that before this CL we wouldn't be able to control hinting on DWrite and DWrite would decide on it by itself, for example from the gasp table? And now Skia sets it explicitly, so is that because there is no big change in behavior?

Patch-set: 2
CC: Dominik Röttsches <6395@5086850b-e599-37d2-8ec8-ff16aeccbf02>