Update patch set 2
Patch Set 2:

Thanks for the clarifications.

> > What do you mean by "because we're not using it as often"? 
> > 
> > Is it correct that before this CL we wouldn't be able to control hinting on DWrite and DWrite would decide on it by itself, for example from the gasp table? And now Skia sets it explicitly, so is that because there is no big change in behavior?
> 
> This shouldn't interact much with anything gasp at all. Hopefully the effect when fully enabled is that when you ask for anti-aliased (but not subpixel anti-aliased) you get anti-aliased (instead of subpixel anti-aliased down sampled). The other part of this is that if the user asks for no hinting then grid fitting will be turned off and the natural symmetric rendering will be used.

But couldn't / shouldn't gasp override the Skia hinting setting? Say setHinting(Full), but gasp says for 7-12 disable hinting, then we should switch it off dependent on gasp?

> Unfortunately, something unexpected is happening with the layout tests and aliased text. It appears some of these must be asking for no hinting or we're accidentally giving them no hinting. I'll be looking into that and perhaps modifying some of this code to take that into account.

Could it be that what you are seeing is what I am intending to rebaseline with the gasp change? The layout tests for sizes outside the minSizeForSubpixel and minSizeForSubpixel antialiasing are combining (for directories other thatn fast/text) non-antialiased drawing with subpixel positioning, which leads to uneven spacing. I am planning to fix this in this large rebaseline.

Patch-set: 2