Alt impl for luminance colorfilter experiment
NOTE: this produces different results from the prev impl.

This is because matrixcolorfilter always operators on unpremul colors,
but this algo wants to operate on premul.

Idea : we create an alternate factory for matrixcolorfilter where the
caller can ask for premul space. That would trivially allow Luma to
be handled via the matrix, and might enable other interesting matrix
effects.

Thoughts?

BUG=skia:

Change-Id: Ib6c988079f770f17e9825287509e96828cba463e