Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> lgtm I guess but I don't really understand how this will fix the ANGLE issue.

ANGLE is ES. When the unit test runs with an ANGLE context the old code would have created a GL second context. The current code will create a GLES context, matching the first context. The new code switches back to creating a GL context.

I still want to try to understand the issue and see if there is a real bug here but I can't repro it and want to get the bots green again in the mean time.

Patch-set: 1