Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (1 comment)
> 
> ... and given reed@'s recent questions about things like color filters in premul/unpremul, it might be nice to have those as independent building blocks, so I'm leaning towards #2.

That sounds better to me. I think the only reason this "ConfigConversionEffect" is a world unto itself is because of the testing it needs to do to select which algorithm to use (or to fall back to sw conversion).

Patch-set: 1
Reviewer: Brian Salomon <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>