Update patch set 1
Patch Set 1:

I'm interested in why there are 9 pieces? (Or why is it 3x3?) Everything else looks good to me.

The only number that I have in my mind is 3 * 4 = 12 intersection points between the rect and the cubic (each rect edge intersects with cubic with at most 3 points, so it's a total of 12 for 4 rect edges).

Patch-set: 1
Reviewer: Yuqian Li <5647@5086850b-e599-37d2-8ec8-ff16aeccbf02>