Update patch set 1
Patch Set 1:

We're now fuzzing the GPU backend with MSAN enabled. MSAN finds a read of uninit in the memcmp in the else block when comparing a single sided stencil setting against a two sided stencil setting. The second face of the single sided stencil settings is never initialized. Trying to think through whether this is OK makes me think that it definitely isn't worth keeping the existing behavior even if it is technically correct as it can't be a significant perf win and is confusing.

Patch-set: 1
Reviewer: Chris Dalton <5633@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Brian Salomon <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>