Update patch set 1
Patch Set 1:

Hal, thank you for doing this.

The next time we're both in the office at the same time, I'd like to show you what I'm working on to improve the documentation. One idea I have that is related to this CL is that associating definitions of things (e.g., how we define a cubic mathematically) is probably best not associated with a specific class member or method since it will be hard to find unless you know what method primarily is associated with that thing.

Secondly, I'd like to propose that member documentation should be the "Cliff's Notes" version which provides just enough to bump-start your memory about what the member is for without requiring that every function fully and completely define parameters and results.

Anyways, this may all make more sense after I show you what I'm up to. Can we hold off on this CL until then?

Patch-set: 1
Reviewer: Cary Clark <5632@5086850b-e599-37d2-8ec8-ff16aeccbf02>