GrTessellator: speed up linearized beziers.
Linerization of quadratic and cubic Beziers was using recursive
midpoint subdivision, which can exceed the number of required points by
up to 2X if the second-last recursion level is just over the flatness
threshold. (This is particularly bad for circular arcs, since all
segments have a similar midpoint distance.)

Instead, compute the points of maximum curvature of the beziers, and
progressively increase the number of segments until the flatness
metric is satisfied at the points of maximum curvature.

This is good for a 15-20% boost on MotionMark Fill Paths, and makes the
performance on Canvas Arcs with the verb limit removed competitive
with the software+upload path renderer.

NOTE: this will potentially affect the GPU results of all GMs with
bezier or conic paths (<= 10 verbs AA, and all MSAA results which
aren't covered by other path renderers).

BUG=skia:

Change-Id: Ia2c22e00e45acec538b15a23c6edcb7a7d4566b7