ok: basic crash handling and stack trace dumps
This seems to work pretty nicely for each engine in the expected case of
few crashes.  The serial and thread engines just dump the first crash
stack to stderr before dying, while the fork engine saves all crashes to
a temporary file, then prints that to stderr once everything's finished.

I'm not settled yet on how best to print what was running when we
crashed, so I'll punt on that for this CL.  TLS is kind of ugly. :/

Change-Id: I0aca66529301b1ad9bd51ec728848817586c606d