ok: fix an assert from unbalanced save/restores
If we're writing .pngs,  the surface passes ownership of its buffer to
an image snapshot, simply marking it as immutable.  Then, when the
surface is destroyed later, it destroys its inner canvas, which restores
its stack to zero, possibly actually doing some drawing in the case of
unbalanced saveLayer()s.  We then hit an assert saying "I just wrote
some pixels but this buffer was marked immutable."

DM didn't find this problem because it's doesn't really use surfaces and
images, just bitmaps.  There's no handoff there and nothing is ever
immutable, so the condition triggering the assert never comes up.

This may actually just be a bug in SkSurface that it doesn't restore to
zero before passing its ownership of the buffer over to the image
snapshot?

Change-Id: Id6d397f534dd1b50219e0d3078c989a4910883a6