Fix SkImage_Gpu::getROPixels
The dstColorSpace is a badly named parameter. It's a hint about where/how
the returned pixels are going to be used. Raster and GPU are meant to
ignore that information - codecs use it to drive our decoding heuristic.

I've adopted that name in several places, although it's quite bad. Maybe
usageColorSpace would be better?

Bug: skia:
Change-Id: I99acaac4bed57ed8acd8433e1c5b39a60c3af09e