Update patch set 3
Patch Set 3:

The N5 try bot disagreed with my claim that no images change, so limiting this to Chrome. It's a small number of GMs on the bot that change by removing this. So I'll want to follow up with whether the bots are running a different OS version than my test device and also whether the workaround could be narrowed.

Patch-set: 3
Reviewer: Brian Salomon <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>