SkCanvas::setLegacy() for the android framework
It only makes sense to wrap legacy canvases with an
SkColorSpaceXformCanvas.

The Android framework has SkBitmaps (with SkColorSpaces) that need
to become legacy SkColorSpaceXformCanvases.

The (ugly) solution proposed here is to force any canvas wrapped by
a SkColorSpaceXformCanvas to be legacy.

A better solution would be a way to share the same pixel ref with
different color spaces.

BUG=skia:

Change-Id: Iabc1d2cbcdaf5d969acd31999d16f492c63138a0