Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Sorry this slipped through my inbox.  This lgtm, but I'm still kind of ¯\_(ツ)_/¯ on transfer function behavior and when we ignore it and how that interacts with premul.

I created SkTransferFnBehavior because there are a few Skia operations that need to be color correct but also need to know whether to handle alpha linearly or in the gamma space.  Ideally, all uses will start internal to Skia and we won't need to expose it in any public APIs.  So far this is used for:
***Premultiplication in SkCodec
***Unpremultiplication in SkImageEncoder
***Premuls and Unpremuls in SkConvertPixels (used by readPixels()/writePixels())

Possible future uses:
***Control blend behavior in SkPictureImageGenerator
***Filtering behavior?

This CL exists because I need to go through Cacherator/Generator to get to SkCodecImGen and SkPictureImGen.

Patch-set: 3
Reviewer: Matt Sarett <5185@5086850b-e599-37d2-8ec8-ff16aeccbf02>