jumper, only ignore red zone on Windows
On Linux and Mac there's always a red zone of 128 bytes of stack space
for us to use without touching the stack pointer.  We'd been generating
stage code as if that's not there because it's not there on Windows.

We have a separate .S file for Windows anyway, so there's no need to
ignore the red zone when we know it's there.

Change-Id: I81a7841020bb8aad68bf35feac851727ef1d0758
Reviewed-on: https://skia-review.googlesource.com/10965
Commit-Queue: Mike Klein <mtklein@chromium.org>
Reviewed-by: Mike Klein <mtklein@chromium.org>
Reviewed-by: Herb Derby <herb@google.com>