jumper, drop Android NDK dependency
We don't _really_ need the Android NDK.  We just need <arm_neon.h>
(which comes from Clang, not the NDK) and a smattering of <stdint.h>
([u]intN_t), <string.h> (memcpy) and <stddef.h> (size_t).

The idea here is solely to make it easier to run build_stages.py.
If this becomes a pain to maintain, I'm happy to go back to the NDK.

Change-Id: Ic6bb287646b6160ac42ac6e4d5290a66a7e92425