Simplify distance computation
Now we have a single, uniform, and simple method of computing all 3 kinds
of simple rrect distances.

It's equivalent with the old way except when the shadow radius is larger
than the rrect radii (but we're still the same when rrect has 0 radii,
which is a simple rect). In the cases that we're different, the new
method is closer to the ground-truth disc blur which will generate a
'+'-shape shadow instead of a 'X'-shape one.

BUG=skia:

Change-Id: I0ec0c5658a0c9ea89edc067e1383e01627979d4e