Set device clip correctly for SkNoDrawCanvas
SkRecorder has some private Skia behavior with the way it handles an
SkNoDrawCanvas that is not exposed.  In particular, it's able to create
an arbitrary SkIRect device clip at construction time.  Without this,
SkNoDrawCanvas(SkIRect) creates a device clip of (width, height) and no
offset, which is confusing behavior.

There's some callers in Chromium (scrollbar painting on mac) that
currently record at some offset and inquire about the device clip.

https://codereview.chromium.org/2768143002 is trying to implement
another version of SkPictureBuilder/SkRecorder and so wants to be able
to use similar SkNoDrawCanvas behavior.

Change-Id: I0ac5f0eef8f0548cefd7354062311667cf448b07