Update patch set 2
Patch Set 2:

(1 comment)

I was always curious when we'd see pictures recorded away from the origin.

Just so I'm clear, this doesn't make a difference for SkPictureRecorder, just your new canvas?  I'm looking at SkPictureRecorder::beginRecording(), which calls SkRecorder::reset(), which calls SkCanvas::resetForNextPicture().

Not that we need to do it now, but do you anticipate wanting to expose resetForNextPicture() in the future?  We found that it was a fair recording speed win to reset the SkCanvas part of the recorder rather than destruct and reconstruct it when recording lots of little pictures.

Patch-set: 2
CC: Mike Klein <5010@5086850b-e599-37d2-8ec8-ff16aeccbf02>