Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> I was always curious when we'd see pictures recorded away from the origin.

Yeah...

You can repro this on Mac by running layout tests, e.g. 

https://storage.googleapis.com/chromium-layout-test-archives/mac_chromium_rel_ng/421575/layout-test-results/results.html

One example, css1/box_properties/border.html: https://storage.googleapis.com/chromium-layout-test-archives/mac_chromium_rel_ng/421575/layout-test-results/css1/box_properties/border-diffs.html
 
Apparently the scrollbars are painted where they are on the page, at a large offset.  That's the only user I've found so far, although plenty of callers believe that they have a rect and not a size.  I saw that comment about the SkNoDrawCanvas(SkIRect) ctor being like "who uses this?!?!".

> Just so I'm clear, this doesn't make a difference for SkPictureRecorder, just your new canvas?  I'm looking at SkPictureRecorder::beginRecording(), which calls SkRecorder::reset(), which calls SkCanvas::resetForNextPicture().

That's exactly it.  SkPictureRecorder::beginRecording(SkRect) does things with SkNoDrawCanvas that I want to do with cc::PaintRecorder::beginRecording to preserve behavior.

I'm going to attempt to wrangle the scrollbar painting code to not do this, but I'm also not 100% confident that this is the only user of this behavior (just the only one that tests yelled about).

> Not that we need to do it now, but do you anticipate wanting to expose resetForNextPicture() in the future?  We found that it was a fair recording speed win to reset the SkCanvas part of the recorder rather than destruct and reconstruct it when recording lots of little pictures.

Yeah, that's a possibility, but not in the short term.  I think unfortunately that Blink really likes to construct its recorders on the stack, and so it'd be a lot of plumbing to reuse the internal SkCanvas.

That would be an easier solution.  I could expose it in SkNoDrawCanvas if you'd like that better.  It'd certainly be less risky of a change.

Patch-set: 2
Reviewer: Adrienne Walker <6595@5086850b-e599-37d2-8ec8-ff16aeccbf02>