Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> non-origin canvas is not well tested for skia, and just now we couldn't recall why we added that (canvas used to always be at 0,0). Therefore it makes us wonder if we can contract and remove that facility in the future.
> 
> Can you make your system work without this? Would you have to pair your canvas with an offset? Could you inject a translate right after creating it and just rely on the width/height of the canvas? Can you inject a clipRect call right after creating it?
> 
> Just trying to understand better the use-case and how intrinsic this needs to be in skia.

Strongly agreed that we shouldn't need this.  (I was also surprised!)  I'm happy to look at modifying callers in Chromium to only use width/height, but that's a larger and longer change.

I had an initial patch that did exactly what you suggested on the Chromium side, but then I realized that if the offset is negative, then the initial device clip rect of 0,0,w,h will not preserve that.  This probably isn't ever used, but painting code is large and it seemed safer to just do what SkPictureRecorder was doing.  It is the case that SkCanvas's qr_clip_bounds already can produce negative offsets and so I can't even assert in Chromium code that the translate/clipRect contains the original rect to verify that no callers are doing something like this.

Patch-set: 2
Reviewer: Adrienne Walker <6595@5086850b-e599-37d2-8ec8-ff16aeccbf02>